requirements.txt
pillow
discord-py-interactions
sympy
